In-depth refactoring, for a better architecture

+++ Rcfw/Inc/encoder.h
@@ -8,7 +8,7 @@
 typedef struct T_ENC_Handle
 {
   char              *name;
-  int16_t            value;
+  int32_t            value;
   bool               invertOnUpdate;
   TIM_HandleTypeDef *timerHandle;
 } T_ENC_Handle;
@@ -19,8 +19,8 @@ void ENC_init(T_ENC_Handle      *p_handle,
               TIM_HandleTypeDef *p_encoderTimerHandle);
 
 void    ENC_reset   (T_ENC_Handle *p_handle                 );
-void    ENC_update  (T_ENC_Handle *p_handle, int16_t p_value);
-int16_t ENC_getCount(T_ENC_Handle *p_handle                 );
+void    ENC_update  (T_ENC_Handle *p_handle, int32_t p_value);
+int32_t ENC_getCount(T_ENC_Handle *p_handle                 );
 void    ENC_logInfo (T_ENC_Handle *p_handle                 );
 
 #endif /* __ENCODER_H */
